refactor(docs): convert routes.py to package + extract lifecycle (#623 PR-A)
PARTIAL #623 — first of an expected 5-PR split.

## What this PR ships

1. `git mv app/docs/routes.py app/docs/routes/__init__.py` — convert the
   module to a package without changing any public import path. Git
   tracks the rename so blame is preserved.
2. Extract `app/docs/routes/_lifecycle.py` with two handlers:
   - `delete_draft_handler` (POST /drafts/{id}/delete)
   - `keep_draft_handler` (POST /drafts/{id}/keep)
3. New `tests/test_docs_routes_registered.py` URL-surface assertion
   that pins all 16 `/drafts*` paths (10 from routes + 6 from
   report_routes) so any future move can detect a dropped route
   immediately. Test is run after each step in the split.
4. Update existing tests in `test_docs_routes.py` and
   `test_archive_warning.py` to patch where the symbol is USED
   (`app.docs.routes._lifecycle.X`) for delete/keep test classes.
   Other test classes still patch `app.docs.routes.X` because their
   handlers stay in `__init__.py` until the next split.

## What's NOT in this PR (deferred to follow-up issues)

- `_upload.py` extraction: `new_draft_page`, `create_draft_handler`,
  upload form helpers (~490 lines).
- `_list.py` extraction: `drafts_list_page`, filter bar, table builders
  (~600 lines).
- `_detail.py` extraction: `draft_detail_page`, status fragment,
  actions fragment, modals (~615 lines).
- `_lifecycle.py` is missing `link_vtk_handler` because it depends on
  `_validate_parent_vtk_fk` (upload) and `_draft_metadata_block`
  (detail). Moving them all together is the next PR.
- `_shared.py` extraction for cross-cutting helpers
  (`_status_tracker`, `_format_elapsed`, `_status_badge`, etc.).

After all splits land, `__init__.py` becomes a registration shim and
no submodule exceeds the DoD's 400-line cap. This first PR shrinks
`__init__.py` from 2611 → 2425 lines — modest but a real start.

## Why partial

The originally-dispatched agent stalled at the 10-minute watchdog
twice on the full-file refactor. Splitting into incremental PRs
makes each step reviewable + testable (URL surface assertion catches
any dropped route per move) and dodges the agent watchdog when we
re-dispatch the remaining splits.

## Verification

- `uv run pytest tests/test_docs_routes_registered.py` — pass
- `uv run pytest tests/` — 2002 passed, 23 skipped (no regressions
  vs the 2001 baseline; the new URL-surface test is the +1)
- `uv run ruff check + ruff format` — clean
- `uv run pyright app/docs/routes/` — 0 errors
- No new migration; no schema change; no behavioural change.

Refs #623.
